placeholder files added for pseudo empty folders

+++ pyprogs/free_space.py
@@ -4,14 +4,20 @@ try:
     os.chdir("./PNG/")
     li = os.listdir()
     for _ in li:
+        if("placeholder.txt" in _):
+            continue
         os.remove(f"./{_}")
     os.chdir("./../SVG/")
     li = os.listdir()
     for _ in li:
+        if("placeholder.txt" in _):
+            continue
         os.remove(f"./{_}")
     os.chdir("./../unplaced/")
     li = os.listdir()
     for _ in li:
+        if("placeholder.txt" in _):
+            continue
         os.remove(f"./{_}")
     print("Success")
 except Exception as e:
